deleted lots of unused files and some beautifulizing
