+++ app/src/main/res/layout/activity_ar_parts.xml
