--- app/src/main/res/layout/activity_begin.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<fragment xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/begin_fragment"
-    android:name="se.stolbygge.stolbygge.BeginFragment"
-    android:layout_width="300dp"
-    android:layout_height="match_parent"
-    android:layout_marginLeft="16dp"
-    android:layout_marginRight="16dp"
-    tools:context=".BeginActivity" />
