+++ app/src/main/res/layout/activity_main.xml
@@ -1,79 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ScrollView
-    android:layout_height="wrap_content"
-    android:layout_width="wrap_content"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_gravity="center_horizontal">
-
-    <LinearLayout
-        android:orientation="vertical"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_gravity="center_horizontal">
-
-        <ImageView
-            android:id="@+id/imageview"
-            android:layout_width="200dp"
-            android:layout_height="200dp"
-            android:layout_marginTop="30dp"
-            android:layout_gravity="center_horizontal"
-            android:background="@drawable/hela_stolen"/>
-
-        <Button
-            android:id="@+id/begin_btn"
-            android:layout_width="130dp"
-            android:layout_height="50dp"
-            android:text="Begin"
-            android:layout_gravity="center_horizontal"
-            android:onClick="onCreateBegin"/>
-
-        <Button
-            android:id="@+id/product_list_btn"
-            android:layout_width="130dp"
-            android:layout_height="50dp"
-            android:text="Product List"
-            android:layout_gravity="center_horizontal"
-            android:onClick="onCreateProductList"/>
-
-        <Button
-            android:id="@+id/instruction_list_btn"
-            android:layout_width="130dp"
-            android:layout_height="50dp"
-            android:text="Instructions"
-            android:layout_gravity="center_horizontal"
-            android:onClick="onCreateInstructionList"/>
-
-        <Button
-            android:id="@+id/result_list_btn"
-            android:layout_width="130dp"
-            android:layout_height="50dp"
-            android:text="Result List"
-            android:layout_gravity="center_horizontal"
-            android:onClick="onCreateResultList"/>
-
-        <Button
-            android:id="@+id/ar_btn"
-            android:layout_width="130dp"
-            android:layout_height="50dp"
-            android:text="AR"
-            android:layout_gravity="center_horizontal"
-            android:onClick="onCreateAR"/>
-
-        <Button
-            android:id="@+id/model_view_btn"
-            android:layout_width="130dp"
-            android:layout_height="50dp"
-            android:text="Model View"
-            android:layout_gravity="center_horizontal"
-            android:onClick="onCreateModelView"/>
-
-        <Button
-            android:id="@+id/ar_instructions_btn"
-            android:layout_width="wrap_content"
-            android:layout_height="50dp"
-            android:text="AR Instructions View"
-            android:layout_gravity="center_horizontal"
-            android:onClick="onCreateARInstructionsView"/>
-
-    </LinearLayout>
-</ScrollView>
+<fragment xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main_fragment"
+    android:name="se.stolbygge.stolbygge.MainFragment"
+    android:layout_width="300dp"
+    android:layout_height="match_parent"
+    android:layout_marginLeft="16dp"
+    android:layout_marginRight="16dp"
+    tools:context=".MainActivity" />
