--- app/src/main/res/layout/activity_missing_parts.xml
@@ -1,12 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MissingPartsActivity">
-
-    <ListView android:id="@+id/missing_list_view"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent">
-    </ListView>
-
-</RelativeLayout>
