+++ app/src/main/res/layout/fragment_ar_instructions.xml
