+++ app/src/main/res/layout/fragment_main.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="se.stolbygge.stolbygge.BeginFragment">
+    tools:context="se.stolbygge.stolbygge.MainFragment">
 
     <!-- Left welcome menu -->
     <LinearLayout
