+++ app/src/main/res/layout/fragment_part_list.xml
