--- app/src/main/res/layout/instructions_activity.xml
@@ -1,33 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="0dp"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context=".MainActivity">
-
-    <ListView
-        android:id="@+id/listview_parts"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent">
-    </ListView>
-
-    <ListView
-        android:id="@+id/listview_steps"
-        android:layout_width="wrap_content"
-        android:layout_height="fill_parent"
-        android:smoothScrollbar="true">
-    </ListView>
-
-    <LinearLayout
-        android:id="@+id/button_list"
-        android:layout_width="wrap_content"
-        android:layout_height="fill_parent"
-        android:orientation="vertical"
-        android:layout_alignParentRight="true"
-        android:padding="0dp">
-
-    </LinearLayout>
-</RelativeLayout>
