--- app/src/main/res/layout/missing_list_item.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <ImageView android:id="@+id/missing_part_image"
-        android:layout_width="100dp"
-        android:layout_height="150dp"
-        android:layout_margin="10dp"
-        android:layout_gravity="center"/>
-
-    <TextView android:id="@+id/missing_part_name"
-        android:layout_width="200dp"
-        android:layout_height="50dp"
-        android:layout_marginBottom="10dp"
-        android:textAppearance="?android:attr/textAppearanceSmall"
-        android:gravity="center"
-        android:layout_gravity="center"/>
-
-</LinearLayout>
