--- app/src/main/res/layout/step_list_item.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="horizontal"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:padding="20dp">
-
-
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="300dp"
-        android:layout_height="300dp"
-        android:layout_weight="1"/>
-
-
-    <RelativeLayout
-        android:layout_width="wrap_content"
-        android:layout_height="fill_parent"
-        android:orientation="vertical"
-        android:paddingTop="100dp"
-        android:paddingBottom="100dp"
-        android:gravity="center_horizontal"
-        android:layout_weight="1">
-
-        <TextView
-            android:id="@+id/textView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceMedium"/>
-
-        <ImageView
-            android:id="@+id/item_checkbox"
-            android:layout_width="200dp"
-            android:layout_height="200dp"
-            android:clickable="true"
-            android:layout_alignParentBottom="true"/>
-    </RelativeLayout>
-
-</LinearLayout>
